+++ src/Feliz.Recoil/Types.fs
@@ -37,6 +37,12 @@ type SelectorGetter =
     [<Emit("$0.get($1)")>]
     member _.get (recoilValue: RecoilValue<'T,ReadWrite>) : 'T = jsNative
 
+[<AutoOpen>]
+module SelectorGetterMagic =
+    type SelectorGetter with
+        [<Emit("$0.get($1)")>]
+        member _.get (recoilValue: RecoilValue<'T,'Mode>) : 'T = jsNative
+
 /// Methods provided in selectors for composing new RecoilValues.
 [<Erase>]
 type SelectorMethods =